refactor: move hero to home & temporarily remove global banner
