+++ src/app/(builder)/[locale]/layout.tsx
@@ -4,8 +4,8 @@ import type { Metadata } from 'next'
 import { notFound } from 'next/navigation'
 import NextTopLoader from 'nextjs-toploader'
 
-import { TopLevelBuilderClientLogic } from '@/app/(builder)/[locale]/topLevelBuilderClientLogic'
 import { TopLevelClientLogic } from '@/app/[locale]/topLevelClientLogic'
+import { TopLevelBuilderClientLogic } from '@/components/app/builder/topLevelBuilderClientLogic'
 import { CookieConsent } from '@/components/app/cookieConsent'
 import { GoogleTagManager } from '@/components/app/googleTagManager'
 import { NavBarGlobalBanner } from '@/components/app/navbarGlobalBanner'
