+++ src/utils/server/builder/models/sections/utils/getSectionContent.ts
@@ -5,12 +5,6 @@ export function getSectionContent(
   sectionModelName: BuilderSectionModelIdentifiers,
   pathname?: string,
 ) {
-  const MyFetchOptions: RequestInit = {
-    method: 'GET',
-    cache: 'no-cache',
-    keepalive: true,
-  }
-
   return builderSDKClient
     .get(sectionModelName, {
       userAttributes: {
@@ -18,7 +12,6 @@ export function getSectionContent(
       },
       prerender: false,
       cachebust: true,
-      fetchOptions: MyFetchOptions,
     })
     .toPromise()
 }
